requirements.txt
autopep8==1.4.3
bottlenose==1.1.8
Django==2.1.3
lxml==4.2.5
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
